improve the speed of list encoding by up to a factor of 2
It is better to traverse the list twice than to explicitly construct a closure
representing the encoding of all its elements, as this avoids a significant
amout of allocation. We gain some additional speed by recursing over builders
only instead of recursing in the 'Put' monad.

Another candidate for the first kind of optimization is 'putSeqOf' and one for
the second kind of optimization is 'putTreeOf'.