Split GSerialize class in two
See the related issue at https://github.com/kolmodin/binary/pull/95,
as this is a known GHC problem. Without this patch, the following
example fails to build due to exhausting the simplifier ticks. With
this patch, it takes less than one second (with GHC 7.10.3)

    {-# LANGUAGE DeriveGeneric #-}
    module A where
    import Data.Serialize
    import GHC.Generics
    data T = T
     () () () () () () () () () ()
     () () () () () () () () () ()
     () () () () () () () () () ()
     () () () () () () () () () ()
     () () () () () () () () () ()
     () () () () () () () () () ()
     () () () () () () () () () ()
     () () () () () () () () () ()
     () () () () () () () () () ()
     () () () () () () () () () ()
     deriving Generic
    instance Serialize T

On the binary-serialise-cbor benchmark suite, this patch has an almost
unbelievable effect in terms of memory usage and compile time for the
cereal instances.

Signed-off-by: Austin Seipp <aseipp@pobox.com>