Fix #68 - memory use of encoding Integers/Naturals
We were building a list of thunks, obtaining the length of the list, pre-pending
a length, then forcing the thunks and serializing each Word8.  The word8s are
fine (slow, but ok) but the forcing of a list of thunks was costing us a lot so
now we compute the length with an integer log instead of construct and traverse
a list, allowing deforestation iiuc.  At any rate, the memory use is way down.

A reasonable test while monitoring the memory use (takes ~5 minutes):

```
import qualified Data.ByteString      as BS
import Data.Serialize

main :: IO ()
main =
 print $ all correct $ [1..1000] ++
                       [2^33 .. 2^33+nr] ++
                       [2^64+589339 .. 2^64 + 600000] ++
                       [2^100-3000 .. 2^100+10] ++
                       [2^big - 4000 .. 2^big + 3000]
 where nr = 1000
       correct :: Integer -> Bool
       correct n =  n == decode' (encode n)
       big = floor (1e5)
```